[back] 스프링 인터셉터를 구현하여 간단한 session 기반 회원가입, 로그인 구현

+++ backend/src/main/java/park_su_park/backend/service/ToDoService.java
@@ -7,11 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 import park_su_park.backend.domain.ToDo;
 import park_su_park.backend.domain.User;
 import park_su_park.backend.dto.requestBody.CreateToDoRequest;
-import park_su_park.backend.dto.responseBody.ToDoDataResponse;
+import park_su_park.backend.dto.responseBody.ToDoData;
 import park_su_park.backend.exception.ForbiddenAccessException;
 import park_su_park.backend.exception.ResourceNotFoundException;
 import park_su_park.backend.repository.ToDoRepository;
-import park_su_park.backend.util.SessionUtil;
+import park_su_park.backend.util.constant.SessionConstant;
+import park_su_park.backend.util.constant.ToDoResponseMessage;
 
 import java.text.MessageFormat;
 
@@ -20,16 +21,15 @@ import java.text.MessageFormat;
 @RequiredArgsConstructor
 public class ToDoService {
 
-    private static final String FORBIDDEN_ACCESS_ACTION = "해당 일정을 수정 및 삭제할 권한이 없습니다";
     private final ToDoRepository toDoRepository;
     private final UserService userService;
 
-    public ToDoDataResponse createToDo(CreateToDoRequest createToDoRequest, HttpSession session) {
-        Long userId = SessionUtil.getUserIdFromSession(session);
-        User user = userService.findUserById(userId);
+    public ToDoData createToDo(CreateToDoRequest createToDoRequest, HttpSession session) {
+        Long userId = (Long) session.getAttribute(SessionConstant.SESSION_USER_ID);
+        User user = userService.findUser(userId);
 
         ToDo toDo = saveToDo(createToDoRequest, user);
-        return ToDoDataResponse.create(toDo);
+        return ToDoData.create(toDo);
     }
 
     private ToDo saveToDo(CreateToDoRequest createToDoRequest, User user) {
@@ -47,33 +47,33 @@ public class ToDoService {
                 ));
     }
 
-    public ToDoDataResponse getToDo(Long toDoId) {
+    public ToDoData getToDoData(Long toDoId) {
         ToDo toDo = findToDo(toDoId);
 
-        return ToDoDataResponse.create(toDo);
+        return ToDoData.create(toDo);
     }
 
-    public ToDoDataResponse updateToDo(CreateToDoRequest updateToDoRequest, Long toDoId, HttpSession session) {
-        ToDo toDo = getAuthorizedToDo(toDoId, session);
+    public ToDoData updateToDo(CreateToDoRequest updateToDoRequest, Long toDoId, HttpSession session) {
+        checkAuthorization(toDoId, session);
+        ToDo toDo = findToDo(toDoId);
         toDo.update(updateToDoRequest);
 
-        return ToDoDataResponse.create(toDo);
+        return ToDoData.create(toDo);
     }
 
     public void deleteToDo(Long toDoId, HttpSession session) {
-        ToDo toDo = getAuthorizedToDo(toDoId, session);
+        checkAuthorization(toDoId, session);
+        ToDo toDo = findToDo(toDoId);
 
         toDoRepository.delete(toDo);
     }
 
-    private ToDo getAuthorizedToDo(Long toDoId, HttpSession session) {
-        Long userId = SessionUtil.getUserIdFromSession(session);
+    private void checkAuthorization(Long toDoId, HttpSession session) {
+        Long userId = (Long) session.getAttribute(SessionConstant.SESSION_USER_ID);
         ToDo toDo = findToDo(toDoId);
 
         if (!toDo.getUser().getId().equals(userId)) {
-            throw new ForbiddenAccessException(FORBIDDEN_ACCESS_ACTION);
+            throw new ForbiddenAccessException(ToDoResponseMessage.TODO_FORBIDDEN_ACCESS);
         }
-
-        return toDo;
     }
 }
